Separate react component and bridges
