+++ src/view/components/overrides/index.js
@@ -1,0 +1,5 @@
+import Overrides from "./overrides";
+
+export { bridge } from "./overridesBridge";
+
+export default Overrides;
